Create export_model function
Move export_model to a library so we can export the full model as part
of the reduction process. This funciton also fixes an issue where U was
appended to W instead of W to U, which causes the states vector for the
full model to have to be in a different format as the reduced model.